fixed stuff pointed out by friends

+++ ft_atoi.c
@@ -6,15 +6,17 @@
 /*   By: eniini <eniini@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/12 10:38:23 by eniini            #+#    #+#             */
-/*   Updated: 2020/07/16 13:30:58 by eniini           ###   ########.fr       */
+/*   Updated: 2020/07/22 00:32:06 by eniini           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int		ft_atoi(const char *str)
 {
-	int result;
-	int negative;
-	int i;
+	long int	result;
+	int 		negative;
+	size_t		i;
 
 	result = 0;
 	negative = 1;
